Another rewrite of patching
Patches are far more complicated than you might think, although at the same time very simple

We work out various attributes of each patch up-front and display that information
as a table. This information is then used to determine how to apply each patch.

To work out if a patch is applicable (and reversible) we must apply it to a copy of the
files that it touches. The --dry-run argument to patch is not enough because if you have
an amalgamated patch that touches the same file in different pre-amalgamated parts then
the fact that it skipped the previous change will often mean that subsequent changes to
the same area will fail.

At the end of the day, I firmly believe that patch quality is important and that there
is an objective scale for the quality of the actual patch blocks (as distinct from the
quality of the changed code, the comments and the commit messages, which are of course
also important). Ideally a patch should apply, apply cleanly, be dry-runnable, be
reversible, not be amalgamated and be at patch level 1

We print out a patch anaylsis at the end of the patching process so the recipe maintainer
is given useful information about which patches could be improved.

If the key comes out as "RAYMD1LOVE" then everything is fine.

Note, this code should be mirrored with git, or git should be used to drive the patching
process with options to automatically clean up patches.