[WIP] Add bisect subcommmand (ugly CLI ATM)
We also implement persistent (between runs) caching for rendered
recipes (for this bisect mode, not in general yet) for which we
add `frozenset_stable()` which calls sys.executable with a constant
PYTHONHASHSEED. We also add `dirhash(path)`.

The status so far is that we use this new class and function to cache
the rendering of recipe directories hashed by the filename, size and
mtime. We could checksum the contents too and probably should.

For an example of using the decorator, see `render_cached()` which
(badly) just wraps `conda_build.api.render()`

An example command-line to test this is:

time \
  conda-build /opt/r/a/python-feedstock -m /opt/r/a/conda_build_config.yaml \
    bisect --start 111 --end 222 --bisect-arg 'test.sh' --bisect-arg 'hello'

Then:

time \
  conda-build /opt/r/a/python-feedstock -m /opt/r/a/conda_build_config.yaml \
    bisect --start 111 --end 222 --bisect-arg 'test.sh' --bisect-arg 'hello'

TODO:

  - [ ] Make -m work, it currently does not and for this to find the right
        variants you must run it from a folder containing the CBC.yaml

  - [ ] Hash the file contents of the recipe. Since we use this for the
        cache filenames its important and we can take the time hit.

This fails because the `conda_build.config.Config` object <class 'MetaData'>
cannot be pickled (because <class 'auxlib._Null'> cannot be picked).

```
Traceback (most recent call last):
  File /opt/Shared.local/asrc/conda-build/conda_build/cli/main_build.py, line 494, in main
    execute(sys.argv[1:])
  File /opt/Shared.local/asrc/conda-build/conda_build/cli/main_build.py, line 483, in execute
    outputs = api.build(args.recipe, post=args.post, test_run_post=args.test_run_post,
  File /opt/Shared.local/asrc/conda-build/conda_build/api.py, line 201, in build
    return build_tree(
  File /opt/Shared.local/asrc/conda-build/conda_build/build.py, line 3020, in build_tree
    metadata_tuples2 = render_recipe_cached(recipe, cfg, no_download_source=False,
  File /opt/Shared.local/asrc/conda-build/conda_build/render.py, line 864, in render_recipe_cached
    return render_recipe(recipe_path, config, no_download_source=no_download_source, variants=variants,
  File /opt/Shared.local/asrc/conda-build/conda_build/utils.py, line 2192, in __call__
    pickle.dump(value, open(pickled, 'wb'))
_pickle.PicklingError: Can't pickle <class 'auxlib._Null'>: import of module 'auxlib' failed
python-BaseException
```