move up tl;dr on separation of build & host
The TL;DR on how to separate build and host sections was placed at the
end of the host sections, somewhat defying its purpose.
This commit moves the basic instructions directly to the build section.
It also tries to remove some duplication and to simplify the flow for a
couple of sentences.