Sort RPMs by version order, add amzn2 and extras/rpm_name
Also fix some dependency issues with amazonlinux2, mainly
perl seems to cyclically depend on lots of perl libs which
then depend on perl.

and encode the hash of __file__ into the pickled file

and write src hashes to speed up rebuilds

and break cycles by tracking already-done packages

and do not move downloaded src archives to the shasummed
name, copy instead, just to speed things up in CDT
generation. We probably do not want this bit finally

and use urllib to quote URLs to avoid issues with +'s
(for example in libsigc++20)

and add --dependency-remove to allow breaking cycles and
add some entires to do this for amazonlinux2

allow empty archives in source.unpack (some RPMs are empty)

use mv/move in source.unpack to avoid permissions errors with
the 'filesystem' RPM packages.