Fix keeping old work when workdir does not match package name
4167 changed the way `compute_build_id` was computed to be based on
the package name. However, it also made the assumption that the old
`work_dir` would be equal to `{croot}/{package name}/work` which
is not true if someone manually passes in a `croot` and a `build_id`
to the `Config` constructor; either manually or as a pass through on
`api.render`.

The result was old work would be lost if the original `work_dir` did not
match the specific string, instead of just looking where the original
work_dir was. This PR fixes the bug and adds a test to make sure the
work is preserved and correctly moved to the new folder.

This is a very niche edge case.