test: Use msumastro=1.1.6 for PyPI skeleton test
msumastro has been updated a couple of days ago and versions >=1.2 now
store the metadata in setup.cfg which is not handled by skeletons.pypi.
It is easy to fix this if we reuse the code from _load_setup_py_data but
we might as well use/advertize newer projects like grayskull instead.