use bash rather than sh on bsd
Many conda recipe build files use shell syntax and commands that are
specific to bash and not POSIX shells.

Given the large number of these that occur it is easier run the build
script in bash than convert them. On Linux and macOS, bash is already
used explicitly.

On FreeBSD, bash is not part of the base system and must be installed
via ports. The install location is /usr/local/bin/bash.