Fix excessive memory use in inspect_linkages_lief
introduced in gh-5228 which replaced binary.name with str(file).
The former returns the path to the binary whereas the latter returns the
contents of the parsed binary, naturally causing high memory usage.
That semantic change was unintended (misread upstream changes).
Since binary's path is no longer part of the object, pass it separately.

Signed-off-by: Marcel Bargull <marcel.bargull@udo.edu>