Fixed a bug in the logic for batching command line arguments
The original algorithm could result in command line calls that were too large. For example, this was causing "The filename or extension is too long" errors on Windows for projects with many files.
Also added the ability to specify the max chunk size length to use via a CONDA_BUILD_MAX_CMD_LEN environment variable.
Fixes #5122